requirements.txt
Quart
discord
discord.py
discord.ui
jishaku
asyncio
discord-ext-menus
aiohttp
datetime
flask
typing
psutil
collection
requests
pathlib
pymongo
wavelink
httpx
discord.ext-context
tasksio
motor
Augmentor
numpy
aiosqlite
gtts
langdetect
python-dotenv
pyttsx3
PyYAML
Requests
openai
duckduckgo-search==6.3.6
colorama
chess
python-dateutil
deep-translator==1.11.4
fastapi
uvicorn
asyncpg
chat_exporter
pyfiglet